Splitting ServerIRC and ServerTelegram out from the core Server module, for clarity
